add resource remapper
(cherry picked from commit 1c9dc6de6633277e17fc0bc7cdada986f239766f)

new api interface

(cherry picked from commit e309574b3667685ccbfe87c26ffe0c53ca6664d3)

lazy rename fix

(cherry picked from commit 66d0eac41d9e6888386ee543f14626bc23f945e0)

create more intuitive name

(cherry picked from commit f952108acef4e97c7cbde54d6414435a94ed82be)

remove lazy, move ResourceRemapper inside OuputConsumerPath

(cherry picked from commit 72efa5a90ca59a24ef3d5c0251c517f52b8d31eb)

fix merge errors

fix rebase
